Support a shell/cat-based command REPL and other fixes.
This required making the JSON Stream handler reopen its input file if
that file is a FIFO.  It will thus run forever and the user must send
SIGINT to terminate the app (as matching expectations from on high).

Fix the wrong capitalization of the FCF (fCF) that crept in.  It must
start with lower case as it is matched against the URL scheme (eg, the
"file" in "file://" matches same in "fileCommandFacility).

Make a "real" (regex) URL parser instead of using the hokey ad-hoc
parser.  This gets broken out into Utils as it may be generally
useful.  A test is added which can be used interactively to check new
URLs.  This parser should handle most, but not all, types of URLs.

Updated docs about the shell-baesd REPL will be available at the
following URL shortly after this push.

  https://brettviren.github.io/dune-daq-repl/ddrepl.html